requirements.txt
Flask
python-pptx
requests
python-docx